+++ src/main/java/by/vyun/bgb/entity/MeetingResultElement.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class MeetingResultElement {
-    Integer userIdTo;
+    Long userIdTo;
     Integer rate;
 }
