+++ src/main/java/by/vyun/bgb/repository/RatingRepo.java
@@ -3,7 +3,7 @@ package by.vyun.bgb.repository;
 import by.vyun.bgb.entity.Rating;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RatingRepo extends JpaRepository<Rating, Integer> {
-    Rating findGameRatingByUserIdAndGameId(int userId, long gameId);
+public interface RatingRepo extends JpaRepository<Rating, Long> {
+    Rating findGameRatingByUserIdAndGameId(Long userId, Long gameId);
 
 }
