+++ src/main/java/by/vyun/bgb/repository/UserRepo.java
@@ -3,8 +3,8 @@ package by.vyun.bgb.repository;
 import by.vyun.bgb.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<User, Integer> {
+public interface UserRepo extends JpaRepository<User, Long> {
     User getFirstByLogin(String login);
-    User getFirstById(Integer id);
+    User getFirstById(Long id);
 
 }
